Added packages to organize classes better for a less cluttered feel
