+++ CS320_Team_Fractal_Website/junit/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/temp/GuessingGameControllerTest.java
@@ -1,12 +1,12 @@
-package edu.ycp.cs320.CS320_Team_Fractal_Website.controller;
+package edu.ycp.cs320.CS320_Team_Fractal_Website.controller.temp;
 
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.GuessingGameController;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.GuessingGame;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.temp.GuessingGameController;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.temp.GuessingGame;
 
 public class GuessingGameControllerTest 
 {
