+++ CS320_Team_Fractal_Website/junit/edu/ycp/cs320/CS320_Team_Fractal_Website/model/MandelbrotTest.java
@@ -5,6 +5,9 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
 
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Location;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Mandelbrot;
+
 public class MandelbrotTest{
 
 	private Mandelbrot model;
