--- CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/LogInController.java
@@ -1,13 +1,0 @@
-package edu.ycp.cs320.CS320_Team_Fractal_Website.controller;
-
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.LogIn;
-
-public class LogInController {
-	private LogIn model;
-	
-	public LogInController(LogIn model){
-		this.model = model;
-	}
-	
-	
-}
