+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/pages/LogInController.java
@@ -1,0 +1,13 @@
+package edu.ycp.cs320.CS320_Team_Fractal_Website.controller.pages;
+
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.pages.LogIn;
+
+public class LogInController {
+	private LogIn model;
+	
+	public LogInController(LogIn model){
+		this.model = model;
+	}
+	
+	
+}
