+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/servlet/MainPageServlet.java
@@ -7,11 +7,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.MandelbrotController;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.SierpinskiController;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.Location;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.Mandelbrot;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.Sierpinski;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.fractal.MandelbrotController;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.fractal.SierpinskiController;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Location;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Mandelbrot;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Sierpinski;
 
 public class MainPageServlet extends HttpServlet
 {
