+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/servlet/ajax/NextGuessAjaxServlet.java
@@ -7,8 +7,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.GuessingGameController;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.GuessingGame;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.temp.GuessingGameController;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.temp.GuessingGame;
 
 public class NextGuessAjaxServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
