+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/servlet/temp/GuessingGameServlet.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.CS320_Team_Fractal_Website.servlet;
+package edu.ycp.cs320.CS320_Team_Fractal_Website.servlet.temp;
 
 import java.io.IOException;
 
@@ -7,8 +7,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.GuessingGameController;
-import edu.ycp.cs320.CS320_Team_Fractal_Website.model.GuessingGame;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.temp.GuessingGameController;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.temp.GuessingGame;
 
 public class GuessingGameServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
